+++ naiserator.yaml
@@ -12,13 +12,6 @@ spec:
   replicas:
     min: 1
     max: 1
-  resources:
-    limits:
-      cpu: 2000m
-      memory: 3000Mi
-    requests:
-      cpu: 500m
-      memory: 600Mi
   liveness:
     path: /tiltak-tilsagnsbrev/internal/healthcheck
   readiness:
